Updated for Version 1.6.1
WARNING: You need to compile against Bukkit and Craftbukkit now(in that
order or you might get ambiguous method errors)

+++ DvZ/src/me/andre111/dvz/item/spell/ItemDamage.java
@@ -71,7 +71,7 @@ public class ItemDamage extends ItemSpell {
 		if(damage>0) {
 			player.damage((double) damage, source);
 		} else {
-			int newHealth = player.getHealth() - damage;
+			double newHealth = player.getHealth() - damage;
 			if(newHealth>player.getMaxHealth()) newHealth = player.getMaxHealth();
 			
 			player.setHealth((double) newHealth);
